refactor: 기존 쉘 api 모킹 로직 분리
